Switch from float to int32_t

+++ FluidNC/src/Maslow/MotorUnit.h
@@ -55,7 +55,7 @@ private:
     int     _numPosErrors                      = 0;  //Keeps track of the number of position errors in a row to detect a stall
     double  _lastPosition                      = 0.0;
     double  _commandPWM                        = 0;  //The last PWM duty cycle sent to the motor
-    double  mostRecentCumulativeEncoderReading = 0;
+    int32_t  mostRecentCumulativeEncoderReading = 0;
     double  encoderReadFailurePrintTime        = millis();
     //unsigned long lastCallGetPos = millis();
